refactor: modularize
