+++ gif-encoder/src/sequential.rs
@@ -1,18 +1,15 @@
 extern crate ffmpeg_next as ffmpeg;
 
 use std::fs::File;
-use std::process::Command;
-use std::time::SystemTime;
 
 use anyhow::{Context, Result, Ok};
-use ffmpeg::format::input;
+
+use crate::core;
 
 pub fn encode_gif(filename: &str) -> Result<()> {
-    let start = SystemTime::now();
-
     // Get video data.
-    let data: Vec<u8> = get_video_data(&filename)?;
-    let (width, height) = get_video_dimensions(&filename)?;
+    let data: Vec<u8> = core::get_video_data(&filename)?;
+    let (width, height) = core::get_video_dimensions(&filename)?;
 
     // Separate the video data into frames.
     let mut data_per_frame: Vec<&[u8]> = data.chunks(
@@ -40,38 +37,5 @@ pub fn encode_gif(filename: &str) -> Result<()> {
         encoder.write_frame(&frame)?;
     }
 
-    let system_duration = start.elapsed().context("Failed to get render time")?;
-    let in_sec = system_duration.as_secs() as f64 + system_duration.subsec_nanos() as f64 * 1e-9;
-    println!("Execution time: {} sec", in_sec);
-
     Ok(())
-}
-
-/**
- * Returns a Vec of bytes for every channel in every pixel in every frame
- * of the video. For example, if the first pixel of the first frame is red,
- * the returned vector's first three elements will be 255, 0, and 0.
-*/
-fn get_video_data(filename: &str) -> Result<Vec<u8>> {
-    let cmd = Command::new("ffmpeg")
-        .args(["-i", filename, "-pix_fmt", "rgb24", "-f", "rawvideo", "-"])
-        .output()
-        .context("Failed to extract frames")?;
-
-    Ok(cmd.stdout.into_iter().collect())
-}
-
-fn get_video_dimensions(filename: &str) -> Result<(usize, usize)> {
-    let input_ctx = input(&filename)?;
-
-    let input = input_ctx
-        .streams()
-        .best(ffmpeg::media::Type::Video)
-        .ok_or(ffmpeg::Error::StreamNotFound)?;
-
-    let decoder = ffmpeg::codec::context::Context::from_parameters(
-        input.parameters()
-    )?.decoder().video()?;
-
-    Ok((decoder.width().try_into()?, decoder.height().try_into()?))
 }
